Configure Paper Theme using an Atomic Strategy #62
Moves conditional platform-specific loading of styles out of the
components `style` prop and into the Theme Config file where it should
be.

Using `Platform.select` in the Theme Configs style definitions, instead
of `Platform.OS` in the individual components `style` prop reduces the
need for the creating of, and importing of multiple Custom Style
properties (E.g., `clogo`, `clogoweb`).

The result is an appropriate separation-of-concerns between
platform-specific styles and component logic in a more streamlined
approach where styles can be extended and re-used by basing "variant"
styles on "base" styles (E.g., `alertContainer` and `msgContainer` both
based on a `container` style which holds the unchanging structural
styles that the container "variants" have in common).

Benefits of employing an Atomic Strategy:

- Reduce Custom Theme Properties
- Cleaner `style` prop for components
- More intuitive Theme Config file
- Variant styles based on Core styles

See #62 for details
Closes #62